SOFTWARE ENGINEERING LABORATORY/Project/OMCS/backend/requirements.txt
absl-py==1.3.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
attrs==21.2.0
autopep8==2.0.1
Babel==2.10.3
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==5.0.1
blinker==1.4
blis==0.7.9
braceexpand==0.1.7
Brlapi==0.8.3
Brotli==1.0.9
build==0.10.0
cachetools==5.0.0
catalogue==2.0.8
certifi==2020.6.20
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
cloudinary==1.32.0
colorama==0.4.4
command-not-found==0.3
confection==0.0.4
contexttimer==0.3.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
dbus-python==1.2.18
debugpy==1.6.3
decorator==4.4.2
decord==0.6.0
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docutils==0.19
docx2pdf==0.1.7
docxcompose==1.4.0
docxtpl==0.16.5
duplicity==0.8.21
einops==0.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
evdev==1.6.0
fairscale==0.4.4
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.8.2
Flask==2.2.2
flatbuffers==22.11.23
fonttools==4.29.1
fs==2.4.12
ftfy==6.1.1
future==0.18.2
gast==0.5.2
gitdb==4.0.10
GitPython==3.1.30
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.12.0
identify==2.5.17
idna==3.3
imageio==2.25.0
importlib-metadata==4.6.4
imutils==0.5.4
iopath==0.1.10
ipykernel==6.16.2
ipython==7.31.1
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.10
jsonlines==3.1.0
jsonschema==4.16.0
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.1
kaggle==1.5.12
keyring==23.5.0
kiwisolver==1.3.2
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libevdev==0.5
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mediapipe==0.9.0
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mpmath==0.0.0
murmurhash==1.0.9
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0
nodeenv==1.7.0
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==2.6.0
plotly==5.13.0
ply==3.11
portalocker==2.7.0
pre-commit==3.0.4
preshed==3.0.8
prometheus-client==0.15.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAutoGUI==0.9.53
pycairo==1.20.1
pycocoevalcap==1.2
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.4
pydeck==0.8.0
PyDrive==1.3.1
PyGetWindow==0.0.9
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.0.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.7.6
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pyproject_hooks==1.0.0
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.28
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-docx==0.8.10
python-dotenv==0.19.2
Python-DS-Assignment-Somya-kr==0.0.1
python-magic==0.4.27
python-slugify==8.0.0
python-xlib==0.32
python3-xlib==0.15
pythran==0.10.0
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyudev==0.22.0
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
readme-renderer==37.3
regex==2022.10.31
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.1
rsa==4.8
salesforce-lavis @ file:///home/somya/Documents/Python_Tkinter_Assignment/LAVIS-main
scikit-image==0.19.3
scipy==1.8.0
SecretStorage==3.3.1
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.97
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
Somya @ file:///home/somya/Documents/SELab/Python_DS_Assignment_Question_02/my_package
soupsieve==2.3.1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.31
sqlparse==0.4.2
srsly==2.4.5
streamlit==1.17.0
sympy==1.9
systemd-python==234
tenacity==8.1.0
terminado==0.17.0
text-unidecode==1.3
thinc==8.1.7
tifffile==2023.2.2
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.26.0
twine==4.0.2
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
validators==0.20.0
virtualenv==20.17.1
wadllib==1.3.6
wasabi==1.1.1
watchdog==2.2.1
wcwidth==0.2.5
webdataset==0.2.31
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
xdg==5
xkit==0.0.0
zipp==1.0.0
